+++ data/genre/src/main/java/com/napzak/market/genre/repositoryimpl/SearchWordGenreRepositoryImpl.kt
@@ -1,8 +1,9 @@
-package com.napzak.market.genre.repository
+package com.napzak.market.genre.repositoryimpl
 
 import com.napzak.market.genre.datasource.SearchWordGenreDataSource
 import com.napzak.market.genre.mapper.toDomain
 import com.napzak.market.genre.model.RecommendedSearchWordGenre
+import com.napzak.market.genre.repository.SearchWordGenreRepository
 import javax.inject.Inject
 
 class SearchWordGenreRepositoryImpl @Inject constructor(
